requirements.txt
google-adk
aiosqlite
